Variavel do tipo double - aula 07 módulo 7

+++ primeiro-programa-java/src/cursojava/executavel/PrimeiraClasseJava.java
@@ -5,15 +5,15 @@ public class PrimeiraClasseJava {
 	/* Main é um método auto executável em Java */
 	public static void main(String[] args) {
 
-		int nota1 = 90;
-		int nota2 = 60;
-		int nota3 = 70;
-		int nota4 = 88;
+		double nota1 = 90;
+		double nota2 = 60;
+		double nota3 = 70;
+		double nota4 = 88;
 		
-		int media = ((nota1 + nota2 + nota3 + nota4)/4);
+		double media = ((nota1 + nota2 + nota3 + nota4)/4);
 		
 		System.out.println("A média das notas é " + media);
-
+		
 	}
 
 }
